requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-generativeai
moviepy
edge-tts
requests
pillow
python-dotenv
apscheduler
pytrends
proglog
